third_party/WebKit/LayoutTests/inspector-protocol/console/console-timestamp-expected.txt

CONSOLE MESSAGE: line 1: testUnique
CONSOLE MESSAGE: line 1: testDouble
CONSOLE MESSAGE: line 1: testDouble
